Added method for backwards compability reasons
There was already a `doGet` method that accepted a map with strings as 2nd argument. As we don't really want to to have use the method with 2 map arguments, added this version of the method. (Would rather switch the names around as I think the header map will be used more, but since the old one existed...)